+++ cmd/ttn-lw-stack/commands/config.go
@@ -27,6 +27,7 @@ import (
 	shared_identityserver "go.thethings.network/lorawan-stack/cmd/internal/shared/identityserver"
 	shared_joinserver "go.thethings.network/lorawan-stack/cmd/internal/shared/joinserver"
 	shared_networkserver "go.thethings.network/lorawan-stack/cmd/internal/shared/networkserver"
+	shared_packetbrokeragent "go.thethings.network/lorawan-stack/cmd/internal/shared/packetbrokeragent"
 	shared_tenantbillingserver "go.thethings.network/lorawan-stack/cmd/internal/shared/tenantbillingserver"
 	"go.thethings.network/lorawan-stack/pkg/applicationserver"
 	conf "go.thethings.network/lorawan-stack/pkg/config"
@@ -39,6 +40,7 @@ import (
 	"go.thethings.network/lorawan-stack/pkg/identityserver"
 	"go.thethings.network/lorawan-stack/pkg/joinserver"
 	"go.thethings.network/lorawan-stack/pkg/networkserver"
+	"go.thethings.network/lorawan-stack/pkg/packetbrokeragent"
 	"go.thethings.network/lorawan-stack/pkg/qrcodegenerator"
 	"go.thethings.network/lorawan-stack/pkg/tenantbillingserver"
 )
@@ -55,6 +57,7 @@ type Config struct {
 	GCS              gatewayconfigurationserver.Config `name:"gcs"`
 	DTC              devicetemplateconverter.Config    `name:"dtc"`
 	QRG              qrcodegenerator.Config            `name:"qrg"`
+	PBA              packetbrokeragent.Config          `name:"pba"`
 
 	DCS deviceclaimingserver.Config `name:"dcs"`
 	CS  cryptoserver.Config         `name:"cs"`
@@ -72,6 +75,7 @@ var DefaultConfig = Config{
 	Console:     shared_console.DefaultConsoleConfig,
 	GCS:         shared_gatewayconfigurationserver.DefaultGatewayConfigurationServerConfig,
 	DTC:         shared_devicetemplateconverter.DefaultDeviceTemplateConverterConfig,
+	PBA:         shared_packetbrokeragent.DefaultPacketBrokerAgentConfig,
 	DCS:         shared_deviceclaimingserver.DefaultDeviceClaimingServerConfig,
 	CS:          shared_cryptoserver.DefaultCryptoServerConfig,
 	TBS:         shared_tenantbillingserver.DefaultTenantBillingServerConfig,